+++ generators/swift/base/src/context/AbstractSwiftGeneratorContext.ts
@@ -300,14 +300,26 @@ export abstract class AbstractSwiftGeneratorContext<
                             this.getSwiftTypeReferenceFromScope(type.keyType, fromSymbol),
                             this.getSwiftTypeReferenceFromScope(type.valueType, fromSymbol)
                         ),
-                    set: () => referencer.referenceAsIsType("JSONValue"),
+                    set: (setType) =>
+                        swift.TypeReference.array(
+                            this.getSwiftTypeReferenceFromScope(
+                                (setType as unknown as { itemType: TypeReference }).itemType,
+                                fromSymbol
+                            )
+                        ),
                     nullable: (ref) =>
                         this.customConfig.nullableAsOptional
                             ? swift.TypeReference.optional(this.getSwiftTypeReferenceFromScope(ref, fromSymbol))
                             : swift.TypeReference.nullable(this.getSwiftTypeReferenceFromScope(ref, fromSymbol)),
                     optional: (ref) =>
                         swift.TypeReference.optional(this.getSwiftTypeReferenceFromScope(ref, fromSymbol)),
-                    list: (ref) => swift.TypeReference.array(this.getSwiftTypeReferenceFromScope(ref, fromSymbol)),
+                    list: (listType) =>
+                        swift.TypeReference.array(
+                            this.getSwiftTypeReferenceFromScope(
+                                (listType as unknown as { itemType: TypeReference }).itemType,
+                                fromSymbol
+                            )
+                        ),
                     _other: () => referencer.referenceAsIsType("JSONValue")
                 });
             case "primitive":