+++ generators/typescript/sdk/generator/src/contexts/type/TypeContextImpl.ts
@@ -355,7 +355,9 @@ export class TypeContextImpl implements TypeContext {
                     return this.needsRequestResponseTypeVariant(typeReference.container.nullable);
                 }
                 if (typeReference.container.type === "list") {
-                    return this.needsRequestResponseTypeVariant(typeReference.container.list);
+                    return this.needsRequestResponseTypeVariant(
+                        (typeReference.container.list as unknown as { itemType: FernIr.TypeReference }).itemType
+                    );
                 }
                 if (typeReference.container.type === "map") {
                     const keyResult = this.needsRequestResponseTypeVariant(typeReference.container.keyType);
@@ -641,11 +643,13 @@ export class TypeContextImpl implements TypeContext {
         return type._visit({
             container: (container) =>
                 container._visit({
-                    list: (value: FernIr.TypeReference) => this.isInline(value),
+                    list: (listType) =>
+                        this.isInline((listType as unknown as { itemType: FernIr.TypeReference }).itemType),
                     map: (value: FernIr.MapType) => this.isInline(value.keyType) || this.isInline(value.valueType),
                     nullable: (value: FernIr.TypeReference) => this.isInline(value),
                     optional: (value: FernIr.TypeReference) => this.isInline(value),
-                    set: (value: FernIr.TypeReference) => this.isInline(value),
+                    set: (setType) =>
+                        this.isInline((setType as unknown as { itemType: FernIr.TypeReference }).itemType),
                     literal: () => false,
                     _other: () => false
                 }),